moving model package location
