+++ app/src/main/java/com/notmiyouji/newsapp/kotlin/model/newsapi/Source.kt
@@ -15,37 +15,17 @@
  *
  */
 
-package com.notmiyouji.newsapp.kotlin.newsapi
+package com.notmiyouji.newsapp.kotlin.model.newsapi
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
 
-class Article {
-    @SerializedName("source")
+class Source {
+    @SerializedName("id")
     @Expose
-    var source: Source? = null
+    var id: String? = null
 
-    @SerializedName("author")
+    @SerializedName("name")
     @Expose
-    var author: String? = null
-
-    @SerializedName("title")
-    @Expose
-    var title: String? = null
-
-    @SerializedName("description")
-    @Expose
-    var description: String? = null
-
-    @SerializedName("url")
-    @Expose
-    var url: String? = null
-
-    @SerializedName("urlToImage")
-    @Expose
-    var urlToImage: String? = null
-
-    @SerializedName("publishedAt")
-    @Expose
-    var publishedAt: String? = null
+    var name: String? = null
 }
