+++ app/src/main/java/com/notmiyouji/newsapp/kotlin/model/user/UserInformation.kt
@@ -1,4 +1,4 @@
-package com.notmiyouji.newsapp.kotlin.model
+package com.notmiyouji.newsapp.kotlin.model.user
 
 import com.google.gson.annotations.Expose
 import com.google.gson.annotations.SerializedName
